Update README.md
This was a project developed in TCSS305 at the University of Washington Tacoma.

It is a simple paint program that provides tools for drawing basic pictures.  I took it upon myself to add save and load features as well as multistep undo and redo functionality.

There were notable design flaws in my original implementation.  If I were to do this today, there would be a greater amount of decoupling tools from the drawing window.  It is nonetheless a functional and reasonably documented basic program.